Fix longstanding race condition in logout CLI command
