+++ libs/arcade-cli/arcade_cli/configure.py
@@ -198,12 +198,10 @@ def configure_client(
         path: Custom path to the MCP client configuration file
     """
     if not from_local and not from_arcade:
-        console.print("[red]Must specify either --from-local or --from-arcade[/red]")
-        raise typer.Exit(1)
+        raise typer.BadParameter("Must specify either --from-local or --from-arcade")
 
     if from_local and from_arcade:
-        console.print("[red]Cannot specify both --from-local and --from-arcade[/red]")
-        raise typer.Exit(1)
+        raise typer.BadParameter("Cannot specify both --from-local and --from-arcade")
 
     # Default server name if not provided
     if not server_name:
@@ -228,6 +226,6 @@ def configure_client(
         else:
             configure_vscode_arcade(server_name, path)
     else:
-        console.print(f"[red]Unknown client: {client}[/red]")
-        console.print("Supported clients: claude, cursor, vscode")
-        raise typer.Exit(1)
+        raise typer.BadParameter(
+            f"Unknown client: {client}. Supported clients: claude, cursor, vscode."
+        )
